Added new files login,sign up and server.php for user management i.e logins and sign ups Added file session_check to validate if user is in sesson else redirect to login. Implemeted session checks and validations in the other files as well. Added function get tasks in get tasks.php to make the get tasks more effient.

+++ components/delete_task.php
@@ -2,9 +2,12 @@
 session_start();
 include "db.php";
 include "notifications.php";
+if (isset($_SESSION['userId'])) {
+    $userState = $_SESSION['userId'];
+}
 if (isset($_POST['deleteTask'])) {
     $task_id = $_POST['task'];
-    $sql = "DELETE FROM tasks where id= '$task_id'";
+    $sql = "DELETE FROM tasks where id= '$task_id' and user_id='$userState' ";
     mysqli_query($conn, $sql);
     addNotification("Task deleted !!!");
     header("location:../index.php");
